+++ server/t01.mjs
@@ -151,7 +151,8 @@ streamList.forEach((stream) => {
   
           });
         } else {
-          console.log('[unknow]');
+          // console.log('[unknow]');
+          // console.log(arr);
 
         }
         // logger.info(arr[0],arr[1],arr[2],arr[3],arr[4],arr[5],arr[6],arr[7],arr[8]);