+++ template/project/webgame4/project/src/component.js
@@ -8,7 +8,8 @@
  *
  ************************************************************************ */
 
-// import m from 'mithril';
+import m from 'mithril';
+
 // import * as comGame from 'component/standalone';
 
 import * as comGame from 'component/gamePIXI';